update 2.17.4 fixed auto purge

+++ Kyoko-MusicBot/bot_files/bot_commands/set_purge.js
@@ -8,8 +8,10 @@ var message_size_delete = setting.message_size_delete
 const commands_setting = require('../bot_setting/commands_setting.json')
 var set_purge = commands_setting.set_purge
 //------------------------------
-exports.run = async (bot,message,msz)=>{
-
+var deleted=false
+//------------------------------
+exports.run = async (bot,message)=>{
+    var msz = 0
     //-----------------------------
     x = fs.readFileSync("./temp/bot_channel_id.json"), err =>{if (err){throw err}}
     var fileback = JSON.parse(x) 
@@ -20,15 +22,22 @@ exports.run = async (bot,message,msz)=>{
     if (!fileback.file_is_writen){
         return
     }else{
-
-        console.log(msz)
-
-        bot_MessChannel.bulkDelete(msz)
-        setTimeout(function() {
-            var embed = new discord.RichEmbed() // message ausgabe
-            .addField("delete the last > "+message_size_delete+" < Messages.","-----------------------------",true )
-            bot_MessChannel.send(embed)
-        }, 1000)
+        if(deleted==false){
+            deleted=true
+            // console.log(msz)            
+            setTimeout(function(){
+                message.channel.fetchMessages({ limit: 100 }).then(messages => {
+                    msz = messages.size   
+                    
+                    bot_MessChannel.bulkDelete(msz)
+                    var embed = new discord.RichEmbed() // message ausgabe
+                    .addField("delete the last > " + msz + " < Messages.","-----------------------------",true )
+                    bot_MessChannel.send(embed)
+                    deleted=false
+                }), err =>{if (err){throw err,console.log(err)}}
+                //----------             
+            }, 2000)
+        }
     }
 }
 //------------------------------
